requirements.txt
amqp==5.2.0
aniso8601==9.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
bcrypt==4.2.0
billiard==4.2.0
blinker==1.5
cachelib==0.9.0
celery==5.4.0
cffi==1.16.0
chardet==5.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Caching==2.3.0
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
fonttools==4.53.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
kombu==5.3.7
lml==0.1.0
MarkupSafe==2.1.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.1
six==1.16.0
SQLAlchemy==2.0.31
texttable==1.7.0
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
WTForms==3.1.2
